Fix android build failing during linking stage with 24.35.30872.22 tag
-ffat-lto-objects flag ensures that the generated object files contain
both the intermediate representation (IR) used for LTO and the regular
machine code.
-fno-fat-lto-objects flag ensures that the generated object files
contain only the intermediate representation (IR) used for LTO, and not
the regular machine code.

With the -ffat-lto-objects flag, android build fails during linking.

Replaced -ffat-lto-objects with -fno-fat-lto-objects as a temporary
workaround to address android build issue.

Signed-off-by: JeevakaPrabu <jeevaka.badrappan@intel.com>